Backend/pyproject.toml
[tool.poetry]
name = "lovelyti-backend"
version = "0.1.0"
description = "lovelyti tv backend"
authors = ["Ahmed Osama <h.ahmedosama@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
pydantic = "^2.5.3"
django-extensions = "^3.2.3"
black = "^23.12.1"
flake8 = "^6.1.0"
djangorestframework-api-key = "^3.0.0"
django-environ = "^0.11.2"
firebase-admin = "^6.3.0"
django-filter = "^23.5"
boto3 = "^1.34.7"
feedparser = "^6.0.11"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
